rename jira types
