+++ front/lib/triggers/built-in-webhooks/jira/components/CreateWebhookJiraConnection.tsx
@@ -14,7 +14,7 @@ import { useEffect, useMemo, useState } from "react";
 
 import type { WebhookCreateFormComponentProps } from "@app/components/triggers/webhook_preset_components";
 import { useWebhookServiceData } from "@app/lib/swr/useWebhookServiceData";
-import type { JiraProjectType } from "@app/lib/triggers/built-in-webhooks/jira/jira_api_types";
+import type { JiraProjectType } from "@app/lib/triggers/built-in-webhooks/jira/types";
 
 export function CreateWebhookJiraConnection({
   owner,
